requirements.txt
pyaudio
streamlit
google-cloud-speech
pydub
python-dotenv
vertexai